Authentication Fix & Flutter SDK Upgrade: Enhancing URL Management:Gsoc (#2838)
* Updated dart sdk to 3.7.2

* Updated ci/cd

* Updated auth method

* Updated auth method

* ci/cd to work with .env

* Did changes according to code rabbit suggestions

* Did changes according to code rabbit suggestions

* Did changes according to suggestions

+++ lib/view_model/pre_auth_view_models/signup_details_view_model.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:graphql_flutter/graphql_flutter.dart';
 import 'package:talawa/constants/app_strings.dart';
 import 'package:talawa/constants/routing_constants.dart';
 import 'package:talawa/enums/enums.dart';
@@ -24,7 +23,7 @@ class SignupDetailsViewModel extends BaseModel {
   late List<Map<String, dynamic>> greeting;
 
   /// Represents information about the selected organization.
-  late OrgInfo? selectedOrganization;
+  late OrgInfo selectedOrganization;
 
   /// Secure local storage instance.
   FlutterSecureStorage secureStorage = const FlutterSecureStorage();
@@ -60,7 +59,7 @@ class SignupDetailsViewModel extends BaseModel {
   ///
   /// **returns**:
   ///   None
-  void initialise(OrgInfo? org) {
+  void initialise(OrgInfo org) {
     selectedOrganization = org;
     // greeting message
     greeting = [
@@ -138,31 +137,15 @@ class SignupDetailsViewModel extends BaseModel {
             ),
           );
           databaseFunctions.init();
-          print("heelo");
-          print(selectedOrganization?.id);
-          final String query;
-          if (selectedOrganization != null) {
-            query = queries.registerUser(
-              firstName.text,
-              lastName.text,
-              email.text,
-              Encryptor.encryptString(
-                password.text,
-              ),
-              selectedOrganization?.id,
-            );
-          } else {
-            query = queries.registerUser(
-              firstName.text,
-              lastName.text,
-              email.text,
-              Encryptor.encryptString(
-                password.text,
-              ),
-              null,
-            );
-          }
-
+          final String query = queries.registerUser(
+            firstName.text,
+            lastName.text,
+            email.text,
+            Encryptor.encryptString(
+              password.text,
+            ),
+            selectedOrganization.id,
+          );
           final result = await databaseFunctions.gqlNonAuthMutation(query);
           navigationService.pop();
           return result;
@@ -175,69 +158,28 @@ class SignupDetailsViewModel extends BaseModel {
             final bool userSaved = await userConfig.updateUser(signedInUser);
             final bool tokenRefreshed = await graphqlConfig.getToken() as bool;
             // if user successfully saved and access token is also generated.
-            if (userSaved && tokenRefreshed) {
-              // if the selected organization userRegistration not required.
-              if (selectedOrganization?.id == '-1') {
-                navigationService.removeAllAndPush(
-                  Routes.mainScreen,
-                  Routes.splashScreen,
-                  arguments: MainScreenArgs(
-                    mainScreenIndex: 0,
-                    fromSignUp: false,
-                  ),
-                );
-                await storingCredentialsInSecureStorage();
-              } else {
-                // the query here will be changes according to if and else once memebership function available
-                final query = queries.joinOrgById(selectedOrganization!.id!);
-                print(query);
-                final QueryResult result =
-                    await databaseFunctions.gqlAuthMutation(
-                  query,
-                );
-                final joinPublicOrganization = result
-                    .data!['joinPublicOrganization'] as Map<String, dynamic>;
-                final List<OrgInfo>? joinedOrg = (joinPublicOrganization[
-                        'joinedOrganizations'] as List<dynamic>?)
-                    ?.map((e) => OrgInfo.fromJson(e as Map<String, dynamic>))
-                    .toList();
-                await userConfig.updateUserJoinedOrg(joinedOrg!);
-                userConfig.saveCurrentOrgInHive(
-                  userConfig.currentUser.joinedOrganizations![0],
-                );
-                navigationService.removeAllAndPush(
-                  Routes.mainScreen,
-                  Routes.splashScreen,
-                  arguments:
-                      MainScreenArgs(mainScreenIndex: 0, fromSignUp: true),
-                );
-                // if (selectedOrganization!.userRegistrationRequired!) {
-
-                // } else {
-                //   final QueryResult result =
-                //       await databaseFunctions.gqlAuthMutation(
-                //     queries.sendMembershipRequest(selectedOrganization!.id!),
-                //   );
-                //   final sendMembershipRequest = result
-                //       .data!['sendMembershipRequest'] as Map<String, dynamic>;
-                //   final OrgInfo membershipRequest = OrgInfo.fromJson(
-                //     sendMembershipRequest['organization']
-                //         as Map<String, dynamic>,
-                //   );
-                //   userConfig.updateUserMemberRequestOrg([membershipRequest]);
-                //   navigationService.pop();
-                //   navigationService.removeAllAndPush(
-                //     Routes.waitingScreen,
-                //     Routes.splashScreen,
-                //   );
-                // }
-                await storingCredentialsInSecureStorage();
-              }
+            if (userSaved &&
+                tokenRefreshed &&
+                userConfig.currentUser.joinedOrganizations != null &&
+                userConfig.currentUser.joinedOrganizations!.isNotEmpty) {
+              userConfig.saveCurrentOrgInHive(
+                userConfig.currentUser.joinedOrganizations![0],
+              );
+              navigationService.removeAllAndPush(
+                Routes.mainScreen,
+                Routes.splashScreen,
+                arguments: MainScreenArgs(mainScreenIndex: 0, fromSignUp: true),
+              );
+              storingCredentialsInSecureStorage();
+            } else {
+              navigationService.showTalawaErrorSnackBar(
+                TalawaErrors.userNotFound,
+                MessageType.error,
+              );
             }
           }
         },
         onActionException: (e) async {
-          print(e);
           navigationService.showTalawaErrorSnackBar(
             'Something went wrong',
             MessageType.error,
